+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to Jodd project are documented here.
 ### Bug fixes
 
 + **mail** - fixed unused `debug` and `timeout`.
++ **mail** - `RFC2822AddressParser` methods `parseToXxx()` now returns `null` for invalid emails.
 + **http** - secure connection was sending `CONNECT` string twice.
 
 ### Breaking changes